refactor: move useFormSync and useCallbackSync
