+++ apps/main/src/llamalend/features/manage-loan/hooks/useMaxRepayTokenValues.ts
@@ -4,11 +4,11 @@ import { useRepayExpectedBorrowed } from '@/llamalend/queries/repay/repay-expect
 import { useRepayIsFull } from '@/llamalend/queries/repay/repay-is-full.query'
 import { useUserState } from '@/llamalend/queries/user'
 import type { RepayFormData, RepayParams } from '@/llamalend/queries/validation/repay.types'
+import { useFormSync } from '@ui-kit/features/forms'
 import type { UseFormReturn } from '@ui-kit/features/forms'
 import { useTokenBalance } from '@ui-kit/hooks/useTokenBalance'
 import { useQueryMinimum } from '@ui-kit/lib'
 import { mapQuery } from '@ui-kit/types/util'
-import { useFormSync } from '@ui-kit/utils/react-form.utils'
 
 export function useMaxRepayTokenValues(
   {
