+++ apps/main/src/llamalend/features/supply/hooks/useMaxDeposit.ts
@@ -2,10 +2,10 @@ import { useMarketVaultMaxDeposit } from '@/llamalend/queries/market'
 import type { DepositForm, DepositParams } from '@/llamalend/queries/validation/supply.validation'
 import type { IChainId as LlamaChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import type { Address } from '@primitives/address.utils'
+import { useFormSync } from '@ui-kit/features/forms'
 import type { UseFormReturn } from '@ui-kit/features/forms'
 import { useTokenBalance } from '@ui-kit/hooks/useTokenBalance'
 import { useQueryMinimum } from '@ui-kit/lib'
-import { useFormSync } from '@ui-kit/utils/react-form.utils'
 
 export function useMaxDepositTokenValues<ChainId extends LlamaChainId>(
   {
