+++ apps/main/src/llamalend/features/supply/hooks/useStakeForm.ts
@@ -10,12 +10,11 @@ import { useFormLowSolvency } from '@/llamalend/widgets/action-card/hooks/useFor
 import type { IChainId as LlamaChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { vestResolver } from '@hookform/resolvers/vest'
 import type { Address } from '@primitives/address.utils'
-import { useForm } from '@ui-kit/features/forms'
+import { useForm, useFormSync } from '@ui-kit/features/forms'
 import { useFormDebounce } from '@ui-kit/hooks/useDebounce'
 import { t } from '@ui-kit/lib/i18n'
 import { LlamaMarketType } from '@ui-kit/types/market'
 import { mapQuery } from '@ui-kit/types/util'
-import { useFormSync } from '@ui-kit/utils/react-form.utils'
 import { useVaultUserBalances } from './useVaultUserBalances'
 
 const userDefaultValues = { stakeAmount: undefined }
