+++ packages/curve-ui-kit/src/features/forms/index.ts
@@ -2,3 +2,5 @@ export * from './useFormContext'
 export * from './useForm'
 export * from './FormProvider'
 export * from './form.types'
+export * from './useFormSync'
+export * from './useCallbackSync'
